Remove rest v1 and Fix test case
