+++ src/main/server/conf/red5.xml
@@ -69,7 +69,7 @@
 		<property name="logLevel" value="${logLevel}"/>   
   	</bean>
 
-	<bean id="ant.media.licence.service" class="#{ T(io.antmedia.rest.BroadcastRestService).isEnterprise() ? 'io.antmedia.enterprise.licence.LicenceService' : 'io.antmedia.licence.CommunityLicenceService' }"
+	<bean id="ant.media.licence.service" class="#{ T(io.antmedia.rest.RestServiceBase).isEnterprise() ? 'io.antmedia.enterprise.licence.LicenceService' : 'io.antmedia.licence.CommunityLicenceService' }"
 	init-method="start">
 		<property name="serverSettings" ref="ant.media.server.settings" />						
 	</bean>
